refactor(core): split agent_loop.rs (5,462 lines) into focused sub-modules
Break the monolithic agent_loop.rs into five sub-modules organized by
concern, improving navigability without changing any public API:

- mod.rs: AgentLoop struct, constructors, blocking turn execution, failover,
  public accessors, nudge injection, and all tests
- types.rs: StreamEvent, AgentResult, AgentLoopConfig, AgentError,
  AgentHooks trait, SubagentSpawner trait, constants
- tools.rs: tool execution (single + parallel), argument parsing,
  fuzzy tool suggestions, policy checking
- streaming.rs: streaming turn execution, SSE chunk collection,
  streamed tool call merging
- compression.rs: context pruning, tool output masking, LLM summarization
- hooks.rs: Lua hook wrappers, shell hook firing, message pushing with
  hook middleware, turn finalization

All 2,740 workspace tests pass. No behavior changes.

Closes #274
